requirements.txt
# Requisitos del Sistema de Gestión de Parqueadero
# Python 3.8+

# ==================== DEPENDENCIAS CORE ====================
# Framework GUI
PyQt5>=5.15.0,<6.0.0

# Conector de Base de Datos MySQL
mysql-connector-python>=8.0.0,<9.0.0

# Hashing de contraseñas (autenticación)
bcrypt>=4.0.0

# ==================== DEPENDENCIAS OPCIONALES ====================
# Exportación de Reportes (Degradación elegante si no están disponibles)
openpyxl>=3.0.0          # Para exportar a Excel
reportlab>=3.6.0         # Para exportar a PDF

# Visualización de Estadísticas
matplotlib>=3.5.0        # Para gráficos estadísticos embebidos

# ==================== RECOMENDADAS PARA PRODUCCIÓN ====================
# Gestión de variables de entorno
python-dotenv>=0.19.0

# ==================== DESARROLLO Y TESTING ====================
# Descomenta estas líneas para desarrollo
# pytest>=7.4.0           # Framework de testing
# pytest-cov>=4.1.0       # Cobertura de tests
# black>=24.3.0           # Formateo de código
# flake8>=7.0.0           # Linting
# isort>=5.13.0           # Ordenar imports
# mypy>=1.8.0             # Type checking
# pre-commit>=3.6.0       # Pre-commit hooks
